refactor: split StageList into smaller components
